pyproject.toml
[project]
name = "csc2701-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-genai>=1.39.1",
    "pandas>=2.3.2",
]

[dependency-groups]
dev = [
    "ipykernel>=6.30.1",
]
